+++ src/Command.ts
@@ -1,0 +1,5 @@
+import {Awaitable} from './Util/Awaitable'
+
+export interface Command <T extends Object> {
+  execute (options: T): Awaitable<void>
+}