+++ src/Handler/Handler.ts
@@ -1,0 +1,5 @@
+import {Program} from '../Program'
+
+export interface Handler {
+  execute (program: Program): Promise<void>
+}